# backend/openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce App
  version: 1.0.0
  description: >-
    A full stack PERN app for selling products online and keeping record of
    customer data.
paths:
...
  /baskets:
    summary: Get all basket data
    description: >-
      This path is used to retrieve all of the baskets data from the postgres
      database.
    post:
      summary: Creates a new basket item
      description: Sends new basket item data to the postgres database
      operationId: post_basket
      requestBody:
        description: A new basket item object
        content:
          application/json:
            schema:
             $ref: '#/components/schemas/Baskets'
      responses:
        default:
          description: Default error sample response
      tags:
        - Baskets
  /baskets/user/:id:
    summary: Get basket data from a user id
    description: This path uses a user ID to collect basket data from the postgres database
    get:
      summary: Gets basket data from a user id
      description: >-
        This path is used to retrieve basket data from a user ID from the
        postgres database
      operationId: get_basket_userId
      responses:
        '200':
          content:
            application/json:
              examples:
                basket_userId:
                  value: |-
                    [
                      {
                        "id": 2,
                        "product_id": 5,
                        "quanity": 1,
                        "user_id": 1
                      },
                      {
                        "id": 3,
                        "product_id": 5,
                        "quanity": 1,
                        "user_id": 1
                      },
                      {
                        "id": 4,
                        "product_id": 6,
                        "quanity": 1,
                        "user_id": 1
                      }]
        default:
          description: Default error sample response
      tags:
        - Baskets
...
components:
  schemas:
...
    Baskets:
      type: object
      properties:
        product_id:
          type: integer
        quanity:
          type: integer
        user_id:
          type: integer
      xml:
        name: Order
...